+++ widgets/tabs.go
@@ -16,7 +16,7 @@ import (
 type TabPane struct {
 	Block
 	TabNames         []string
-	ActiveTabIndex   uint
+	ActiveTabIndex   int
 	ActiveTabStyle   Style
 	InactiveTabStyle Style
 }
@@ -37,7 +37,7 @@ func (self *TabPane) FocusLeft() {
 }
 
 func (self *TabPane) FocusRight() {
-	if self.ActiveTabIndex < uint(len(self.TabNames)-1) {
+	if self.ActiveTabIndex < len(self.TabNames)-1 {
 		self.ActiveTabIndex++
 	}
 }
@@ -48,7 +48,7 @@ func (self *TabPane) Draw(buf *Buffer) {
 	xCoordinate := self.Inner.Min.X
 	for i, name := range self.TabNames {
 		ColorPair := self.InactiveTabStyle
-		if uint(i) == self.ActiveTabIndex {
+		if i == self.ActiveTabIndex {
 			ColorPair = self.ActiveTabStyle
 		}
 		buf.SetString(
